requirements_step4.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
joblib==1.4.2
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
